Put standalone .vmdk image in a .zip archive
This allows the standalone image to be downloaded as quickly as the .ova
image. (The text notes that one might one to use the standalone image because
a compressed .vmdk might not be suitable for some situations or
virtualization platforms, but that's no reason the uncompressed .vmdk can't
be wrapped in a .zip for download. Just unzip it and you'll have an
uncompressed .vmdk ready for use with your virtualization software.)